+++ packages/portalnetwork/src/client/provider.ts
@@ -1,6 +1,6 @@
 import { ethers } from 'ethers'
 import { addRLPSerializedBlock, HistoryProtocol } from '../subprotocols/index.js'
-import { ProtocolId } from '../subprotocols/types.js'
+import { ProtocolId } from '../types.js'
 import { toHexString } from '../util/discv5.js'
 import {
   ethJsBlockToEthersBlock,
