+++ packages/portalnetwork/src/index.ts
@@ -5,3 +5,4 @@ export * from './util/index.js'
 export * from './transports/index.js'
 export * from './subprotocols/index.js'
 export * from './client/index.js'
+export * from './types.js'
