+++ packages/portalnetwork/src/types.ts
@@ -1,0 +1,11 @@
+// subprotocol IDs
+export enum ProtocolId {
+  StateNetwork = '0x500a',
+  HistoryNetwork = '0x500b',
+  TxGossipNetwork = '0x500c',
+  HeaderGossipNetwork = '0x500d',
+  CanonicalIndicesNetwork = '0x500e',
+  BeaconLightClientNetwork = '0x501a',
+  UTPNetwork = '0x757470',
+  Rendezvous = '0x72656e',
+}
