+++ packages/portalnetwork/vitest.config.unit.ts
@@ -3,7 +3,7 @@ import { defineConfig } from 'vitest/config'
 export default defineConfig({
   test: {
     silent: true,
-    exclude: ['test/integration', 'test/client/provider.spec.ts', 'test/subprotocols/history/types.spec.ts'],
+    exclude: ['test/integration'],
     testTimeout: 180000,
   },
 })
